It's now 64bits compatible it seems.

+++ src/game/FileCRCList.h
@@ -16,17 +16,18 @@
 #ifndef _FILECRCLIST_H
 #define _FILECRCLIST_H
 
+#include <stdint.h>
 #include "CStr.h"
 #include "Trees.h"
 
 class FileCRCList : public LinklistBase<FileCRCList> {
 public:
 	CStr filename;
-	unsigned long crc;
+	uint32_t crc;
 	int size;
 	FileCRCList() {filename = ""; crc = 0;}
-//	FileCRCList(char *file, unsigned long hash) {filename = file; crc = hash;}
-	FileCRCList(const char *file, unsigned long hash, int len) {filename = file; crc = hash; size = len;}
+//	FileCRCList(char *file, uint32_t hash) {filename = file; crc = hash;}
+	FileCRCList(const char *file, uint32_t hash, int len) {filename = file; crc = hash; size = len;}
 
 	bool FileCRC(const char *file);
