+++ src/game/Reg.h
@@ -38,6 +38,7 @@
 
 #include <string>
 #include <unordered_map>
+#include <cstdint>
 
 class Registry
 {
@@ -48,8 +49,8 @@ public:
 
 	~Registry();
 
-	void WriteDword(const char *name, unsigned long val) { WriteValue(name, val); }
-	void ReadDword(const char *name, unsigned long *val) const { ReadValue(name, *val); }
+	void WriteDword(const char *name, uint32_t val) { WriteValue(name, val); }
+	void ReadDword(const char *name, uint32_t *val) const { ReadValue(name, *val); }
 	void ReadDword(const char *name, int *val) const { ReadValue(name, *val); }
 	void ReadDword(const char *name, bool *val) const { ReadValue(name, *val); }
 	void WriteFloat(const char *name, float val) { WriteValue(name, val); }
@@ -66,7 +67,7 @@ public:
 			Cast(it->second, val);
 	}
 
-	void Cast(const std::string& in, unsigned long& out) const { out = std::stoul(in); }
+	void Cast(const std::string& in, uint32_t& out) const { out = std::stoul(in); }
 	void Cast(const std::string& in, int& out) const { out = std::stoi(in); }
 	void Cast(const std::string& in, bool& out) const { out = !in.empty() && in[0] == 't' || in[0] == 'T' || in[0] == '1'; }
 	void Cast(const std::string& in, float& out) const { out = std::stof(in); }
